revert some changes in constants directory
